sorting_system/requirements-ci.txt
# requirements-ci.txt
# Requirements for CI/CD environment

# Core Django framework
Django>=5.2.1,<6.0
djangorestframework>=3.14.0

# Database
# Using SQLite for CI (default Django database)
# psycopg2-binary>=2.9.0  # Uncomment for PostgreSQL in production

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Machine Learning
torch==2.0.1+cpu
torchvision==0.15.2+cpu
timm>=0.9.0
ultralytics>=8.0.0

# Web server
gunicorn>=21.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-django>=4.5.0

# Production dependencies (commented out for CI speed)
# tensorflow>=2.13.0  # Optional: For TensorFlow models
# scikit-learn>=1.3.0  # Optional: For traditional ML models